move pure python collection serializers out
