+++ python/pyfory/struct.py
@@ -14,6 +14,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+
 import dataclasses
 import datetime
 import enum
